[BONFIRE] refactored the MemoryMonitorExecutorExtension into its own class
